+++ R/ipredict.R
@@ -1,0 +1,3 @@
+ipredict <-
+function (object, ...)
+    UseMethod("ipredict")